Move test doubles into independent files.
refs #125
