+++ tests/TestApp/Model/Filter/TestFilter.php
@@ -1,0 +1,12 @@
+<?php
+namespace Search\Test\TestApp\Model\Filter;
+
+use Search\Model\Filter\Base;
+
+class TestFilter extends Base
+{
+
+    public function process()
+    {
+    }
+}
