+++ tests/TestApp/Model/Table/ArticlesTable.php
@@ -1,0 +1,20 @@
+<?php
+namespace Search\Test\TestApp\Model\Table;
+
+use Cake\ORM\Table;
+use Search\Manager;
+
+class ArticlesTable extends Table
+{
+
+    public function searchConfiguration()
+    {
+        $manager = new Manager($this);
+
+        return $manager
+            ->value('foo')
+            ->like('search', ['filterEmpty' => true])
+            ->value('baz')
+            ->value('group', ['field' => 'Articles.group']);
+    }
+}
